separate img input from caption entry
