+++ client/features/camera/Camera.jsx
@@ -1,46 +1,22 @@
 import React from 'react'
-import { useSelector, useDispatch } from 'react-redux'
+import { useDispatch } from 'react-redux'
+import { useHistory } from 'react-router-dom'
 
-import { setImg, setCaption } from './camProtoSlice'
+import { setImg } from './camProtoSlice'
 
 function Camera () {
-  const { img, caption } = useSelector(state => state.camera)
   const dispatch = useDispatch()
+  const history = useHistory()
 
   function addImg (e) {
     const img = e.target.files[0]
     const imgURL = URL.createObjectURL(img)
     dispatch(setImg(imgURL))
-  }
-
-  function addCaption (e) {
-    const { value } = e.target
-    dispatch(setCaption(value))
-  }
-
-  function resetForm () {
-    dispatch(setImg(''))
-    dispatch(setCaption(''))
-  }
-
-  function submitForm () {
-    console.log(img)
-    console.log(caption)
+    history.push('/caption')
   }
 
   return (
-    <form onSubmit={submitForm}>
-      <button onClick={resetForm}>Back</button>
-      {
-        img
-          ? (<>
-            <img src={img}/>
-            <input type='text' placeholder='caption' value={caption} onChange={addCaption}/>
-            <button type='submit'>Submit</button>
-          </>)
-          : <input type='file' id='imageFile' capture='environment' accept='image/*' onChange={addImg}/>
-      }
-    </form>
+    <input type='file' id='imageFile' capture='environment' accept='image/*' onChange={addImg}/>
   )
 }
